Splitting tile providers back into individual files
