+++ modules/globebrowsing/CMakeLists.txt
@@ -57,9 +57,19 @@ set(HEADER_FILES
   src/skirtedgrid.h
   src/tileindex.h
   src/tileloadjob.h
-  src/tileprovider.h
   src/tiletextureinitdata.h
   src/timequantizer.h
+  src/tileprovider/defaulttileprovider.h
+  src/tileprovider/imagesequencetileprovider.h
+  src/tileprovider/interpolatetileprovider.h
+  src/tileprovider/singleimagetileprovider.h
+  src/tileprovider/sizereferencetileprovider.h
+  src/tileprovider/temporaltileprovider.h
+  src/tileprovider/texttileprovider.h
+  src/tileprovider/tileindextileprovider.h
+  src/tileprovider/tileprovider.h
+  src/tileprovider/tileproviderbyindex.h
+  src/tileprovider/tileproviderbylevel.h
 )
 
 set(SOURCE_FILES
@@ -87,9 +97,19 @@ set(SOURCE_FILES
   src/skirtedgrid.cpp
   src/tileindex.cpp
   src/tileloadjob.cpp
-  src/tileprovider.cpp
   src/tiletextureinitdata.cpp
   src/timequantizer.cpp
+  src/tileprovider/defaulttileprovider.cpp
+  src/tileprovider/imagesequencetileprovider.cpp
+  src/tileprovider/interpolatetileprovider.cpp
+  src/tileprovider/singleimagetileprovider.cpp
+  src/tileprovider/sizereferencetileprovider.cpp
+  src/tileprovider/temporaltileprovider.cpp
+  src/tileprovider/texttileprovider.cpp
+  src/tileprovider/tileindextileprovider.cpp
+  src/tileprovider/tileprovider.cpp
+  src/tileprovider/tileproviderbyindex.cpp
+  src/tileprovider/tileproviderbylevel.cpp
 )
 source_group("Source Files" FILES ${SOURCE_FILES})
