+++ modules/globebrowsing/src/layer.cpp
@@ -286,13 +286,13 @@ Layer::Layer(layergroupid::GroupID id, const ghoul::Dictionary& layerDict,
 
     _reset.onChange([&]() {
         if (_tileProvider) {
-            tileprovider::reset(*_tileProvider);
+            _tileProvider->reset();
         }
     });
 
     _remove.onChange([&]() {
         if (_tileProvider) {
-            tileprovider::reset(*_tileProvider);
+            _tileProvider->reset();
             _parent.deleteLayer(identifier());
         }
     });
@@ -360,13 +360,13 @@ void Layer::initialize() {
     ZoneScoped
 
     if (_tileProvider) {
-        tileprovider::initialize(*_tileProvider);
+        _tileProvider->initialize();
     }
 }
 
 void Layer::deinitialize() {
     if (_tileProvider) {
-        tileprovider::deinitialize(*_tileProvider);
+        _tileProvider->deinitialize();
     }
 }
 
@@ -374,7 +374,7 @@ ChunkTilePile Layer::chunkTilePile(const TileIndex& tileIndex, int pileSize) con
     ZoneScoped
 
     if (_tileProvider) {
-        return tileprovider::chunkTilePile(*_tileProvider, tileIndex, pileSize);
+        return _tileProvider->chunkTilePile(tileIndex, pileSize);
     }
     else {
         ChunkTilePile chunkTilePile;
@@ -390,7 +390,7 @@ ChunkTilePile Layer::chunkTilePile(const TileIndex& tileIndex, int pileSize) con
 
 Tile::Status Layer::tileStatus(const TileIndex& index) const {
     return _tileProvider ?
-        tileprovider::tileStatus(*_tileProvider, index) :
+        _tileProvider->tileStatus(index) :
         Tile::Status::Unavailable;
 }
 
@@ -404,7 +404,7 @@ layergroupid::BlendModeID Layer::blendMode() const {
 
 TileDepthTransform Layer::depthTransform() const {
     return _tileProvider ?
-        tileprovider::depthTransform(*_tileProvider) :
+        _tileProvider->depthTransform() :
         TileDepthTransform{ 1.f, 0.f };
 }
 
@@ -416,7 +416,7 @@ bool Layer::enabled() const {
     return _enabled;
 }
 
-tileprovider::TileProvider* Layer::tileProvider() const {
+TileProvider* Layer::tileProvider() const {
     return _tileProvider.get();
 }
 
@@ -440,7 +440,7 @@ void Layer::update() {
     ZoneScoped
 
     if (_tileProvider) {
-        tileprovider::update(*_tileProvider);
+        _tileProvider->update();
     }
 }
 
@@ -483,7 +483,7 @@ void Layer::initializeBasedOnType(layergroupid::TypeID id, ghoul::Dictionary ini
                 std::string name = initDict.value<std::string>(KeyName);
                 LDEBUG("Initializing tile provider for layer: '" + name + "'");
             }
-            _tileProvider = tileprovider::createFromDictionary(id, std::move(initDict));
+            _tileProvider = TileProvider::createFromDictionary(id, std::move(initDict));
             break;
         }
         case layergroupid::TypeID::SolidColor: {
