+++ modules/globebrowsing/src/tileprovider/interpolatetileprovider.h
@@ -22,59 +22,41 @@
  * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
  ****************************************************************************************/
 
-#ifndef __OPENSPACE_MODULE_GLOBEBROWSING___LAYERGROUP___H__
-#define __OPENSPACE_MODULE_GLOBEBROWSING___LAYERGROUP___H__
+#ifndef __OPENSPACE_MODULE_GLOBEBROWSING___TILEPROVIDER__INTERPOLATEPROVIDER___H__
+#define __OPENSPACE_MODULE_GLOBEBROWSING___TILEPROVIDER__INTERPOLATEPROVIDER___H__
 
-#include <openspace/properties/propertyowner.h>
+#include <modules/globebrowsing/src/tileprovider/tileprovider.h>
 
-#include <modules/globebrowsing/src/layergroupid.h>
-#include <openspace/properties/scalar/boolproperty.h>
+#include <modules/globebrowsing/src/tileprovider/singleimagetileprovider.h>
 
 namespace openspace::globebrowsing {
 
-class Layer;
-struct TileProvider;
+struct InterpolateTileProvider : public TileProvider {
+    InterpolateTileProvider(const ghoul::Dictionary&);
+    virtual ~InterpolateTileProvider();
 
-/**
- * Convenience class for dealing with multiple <code>Layer</code>s.
- */
-struct LayerGroup : public properties::PropertyOwner {
-    LayerGroup(layergroupid::GroupID id);
+    Tile tile(const TileIndex& tileIndex) override final;
+    Tile::Status tileStatus(const TileIndex& index) override final;
+    TileDepthTransform depthTransform() override final;
+    void update() override final;
+    void reset() override final;
+    int maxLevel() override final;
+    float noDataValueAsFloat() override final;
 
-    void setLayersFromDict(const ghoul::Dictionary& dict);
-
-    void initialize();
-    void deinitialize();
-
-    /// Updates all layers tile providers within this group
-    void update();
-
-    Layer* addLayer(const ghoul::Dictionary& layerDict);
-    void deleteLayer(const std::string& layerName);
-    void moveLayers(int oldPosition, int newPosition);
-
-    /// @returns const vector of all layers
-    std::vector<Layer*> layers() const;
-
-    /// @returns const vector of all active layers
-    const std::vector<Layer*>& activeLayers() const;
-
-    /// @returns the size of the pile to be used in rendering of this layer
-    int pileSize() const;
-
-    bool layerBlendingEnabled() const;
-
-    void onChange(std::function<void(Layer*)> callback);
-
-private:
-    const layergroupid::GroupID _groupId;
-    std::vector<std::unique_ptr<Layer>> _layers;
-    std::vector<Layer*> _activeLayers;
-
-    properties::BoolProperty _levelBlendingEnabled;
-    std::function<void(Layer*)> _onChangeCallback;
+    TileProvider* before = nullptr;
+    TileProvider* t1 = nullptr;
+    TileProvider* t2 = nullptr;
+    TileProvider* future = nullptr;
+    float factor = 1.f;
+    GLuint vaoQuad = 0;
+    GLuint vboQuad = 0;
+    GLuint fbo = 0;
+    std::string colormap;
+    std::unique_ptr<ghoul::opengl::ProgramObject> shaderProgram;
+    std::unique_ptr<SingleImageProvider> singleImageProvider;
+    cache::MemoryAwareTileCache* tileCache = nullptr;
 };
 
 } // namespace openspace::globebrowsing
 
-#endif // __OPENSPACE_MODULE_GLOBEBROWSING___LAYERGROUP___H__
+#endif // __OPENSPACE_MODULE_GLOBEBROWSING___TILEPROVIDER__INTERPOLATEPROVIDER___H__
