+++ modules/globebrowsing/src/tileprovider/singleimagetileprovider.cpp
@@ -22,59 +22,72 @@
  * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
  ****************************************************************************************/
 
-#ifndef __OPENSPACE_MODULE_GLOBEBROWSING___LAYERGROUP___H__
-#define __OPENSPACE_MODULE_GLOBEBROWSING___LAYERGROUP___H__
+#include <modules/globebrowsing/src/tileprovider/singleimagetileprovider.h>
 
-#include <openspace/properties/propertyowner.h>
+#include <ghoul/io/texture/texturereader.h>
 
-#include <modules/globebrowsing/src/layergroupid.h>
-#include <openspace/properties/scalar/boolproperty.h>
+namespace {
+    constexpr const char* KeyFilePath = "FilePath";
+    
+    constexpr openspace::properties::Property::PropertyInfo FilePathInfo = {
+        "FilePath",
+        "File Path",
+        "The file path that is used for this image provider. The file must point to an "
+        "image that is then loaded and used for all tiles."
+    };
+} // namespace
 
 namespace openspace::globebrowsing {
 
-class Layer;
-struct TileProvider;
+SingleImageProvider::SingleImageProvider(const ghoul::Dictionary& dictionary)
+    : filePath(FilePathInfo)
+{
+    ZoneScoped
 
-/**
- * Convenience class for dealing with multiple <code>Layer</code>s.
- */
-struct LayerGroup : public properties::PropertyOwner {
-    LayerGroup(layergroupid::GroupID id);
+    filePath = dictionary.value<std::string>(KeyFilePath);
+    addProperty(filePath);
 
-    void setLayersFromDict(const ghoul::Dictionary& dict);
+    reset();
+}
 
-    void initialize();
-    void deinitialize();
+Tile SingleImageProvider::tile(const TileIndex&) {
+    ZoneScoped
+    return ttile;
+}
 
-    /// Updates all layers tile providers within this group
-    void update();
+Tile::Status SingleImageProvider::tileStatus(const TileIndex&) {
+    return ttile.status;
+}
 
-    Layer* addLayer(const ghoul::Dictionary& layerDict);
-    void deleteLayer(const std::string& layerName);
-    void moveLayers(int oldPosition, int newPosition);
+TileDepthTransform SingleImageProvider::depthTransform() {
+    return { 0.f, 1.f };
+}
 
-    /// @returns const vector of all layers
-    std::vector<Layer*> layers() const;
+void SingleImageProvider::update() {}
 
-    /// @returns const vector of all active layers
-    const std::vector<Layer*>& activeLayers() const;
+void SingleImageProvider::reset() {
+    if (filePath.value().empty()) {
+        return;
+    }
+    tileTexture = ghoul::io::TextureReader::ref().loadTexture(filePath, 2);
+    if (!tileTexture) {
+        throw ghoul::RuntimeError(
+            fmt::format("Unable to load texture '{}'", filePath.value())
+        );
+    }
+    Tile::Status tileStatus = Tile::Status::OK;
 
-    /// @returns the size of the pile to be used in rendering of this layer
-    int pileSize() const;
+    tileTexture->uploadTexture();
+    tileTexture->setFilter(ghoul::opengl::Texture::FilterMode::AnisotropicMipMap);
+    ttile = Tile{ tileTexture.get(), std::nullopt, tileStatus };
+}
 
-    bool layerBlendingEnabled() const;
+int SingleImageProvider::maxLevel() {
+    return 1337; // unlimited
+}
 
-    void onChange(std::function<void(Layer*)> callback);
-
-private:
-    const layergroupid::GroupID _groupId;
-    std::vector<std::unique_ptr<Layer>> _layers;
-    std::vector<Layer*> _activeLayers;
-
-    properties::BoolProperty _levelBlendingEnabled;
-    std::function<void(Layer*)> _onChangeCallback;
-};
+float SingleImageProvider::noDataValueAsFloat() {
+    return std::numeric_limits<float>::min();
+}
 
 } // namespace openspace::globebrowsing
-
-#endif // __OPENSPACE_MODULE_GLOBEBROWSING___LAYERGROUP___H__
